add initial tests to full backend test runner
It's important to keep running the mock tests: If we can't pass the mock
tests, we can't possibly pass the integration tests. Or, worse, if we
fail the mock tests and pass the integration tests, it may mean our mock
is bad.

+++ share/run_backend_tests.sh
@@ -1,4 +1,11 @@
 
+echo "-- Running Test (mock) tests"
+prove t
+if [ $? -gt 0 ]; then
+    echo "Failed Test (mock) tests"
+    exit 1
+fi
+
 echo "-- Running SQLite tests"
 rm -f /tmp/yancy_sqlite3.db
 sqlite3 /tmp/yancy_sqlite3.db < t/share/sqlite.sql
